ResolveZA/Backend/requirements.txt

# --- Web Framework ---
fastapi==0.111.0
uvicorn[standard]==0.30.1

# --- Database ---
sqlalchemy==2.0.30         # ORM with async support
pyodbc==5.1.0              # SQL Server ODBC driver
alembic==1.13.1            # Database migrations

# --- Validation ---
pydantic==2.7.1            # Request/response schemas
pydantic-settings==2.2.1   # Environment variable config

# --- Authentication ---
python-jose[cryptography]==3.3.0   # JWT tokens
passlib[bcrypt]==1.7.4             # Password hashing

# --- Async Jobs ---
celery==5.4.0
redis==5.0.4

# --- HTTP ---
httpx==0.27.0              # Async HTTP client (used in tests too)

# --- Utilities ---
python-multipart==0.0.9    # Form data / file uploads
structlog==24.1.0          # Structured logging (JSON output)

# --- Testing ---
pytest==8.2.2
pytest-asyncio==0.23.7
pytest-cov==5.0.0

# --- Linting ---
ruff==0.4.7